# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API v2",
        "contact": {
            "email": "ybier@esme-solutions.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/points": {
            "get": {
                "tags": [
                    "Points"
                ],
                "summary": "Get a list of points",
                "operationId": "ec8a34f7b756bf02251422a3d5ca1688",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Points"
                ],
                "summary": "Create a point",
                "operationId": "69fb93a238e661b25388836e4bbc1544",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ben Site"
                                    },
                                    "site_id": {
                                        "description": "ID of the site",
                                        "type": "string",
                                        "example": 1
                                    },
                                    "meta": {
                                        "type": "object",
                                        "example": [
                                            "deserunt",
                                            "amet",
                                            "sit",
                                            "illo",
                                            "labore"
                                        ]
                                    },
                                    "status": {
                                        "description": "Status of the point. 1 or 0",
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/points/{id}": {
            "get": {
                "tags": [
                    "Points"
                ],
                "summary": "Show details of a point",
                "operationId": "1e6e11201bccd533b64eb408bd01dd5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the point to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Points"
                ],
                "summary": "Update a point",
                "operationId": "307cb07784f48ef7fb1f01c1e5424944",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the point to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Ben Site"
                                            },
                                            "site_id": {
                                                "description": "ID of the site",
                                                "type": "string",
                                                "example": 1
                                            },
                                            "meta": {
                                                "type": "object",
                                                "example": [
                                                    "deserunt",
                                                    "amet",
                                                    "sit",
                                                    "illo",
                                                    "labore"
                                                ]
                                            },
                                            "status": {
                                                "description": "Status of the point. 1 or 0",
                                                "type": "string",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Points"
                ],
                "summary": "Delete a point",
                "operationId": "d1b54e98d4b30c7a581f6b8b0677fda7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the point to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "clientCredentials": {
                        "authorizationUrl": "http://powerbat-laravel/oauth/authorize",
                        "tokenUrl": "http://powerbat-laravel/oauth/token",
                        "refreshUrl": "http://powerbat-laravel/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}